Moved collectQuoteFiles into QuoteForm component
